+++ src/index.ts
@@ -1,22 +1,20 @@
-import {
-  concatBytes,
-  toBytes,
-  parseHexByte,
-  numberToHex,
-  hexToBytes,
-  bytesToHex,
-  utf8ToBytes,
-} from './utils'
-import { Decoded, Input, List } from './types'
+export type Input = string | number | bigint | Uint8Array | List | null
 
-// Types exported for easy use
-export { Decoded, Input, List }
+// Use interface extension instead of type alias to
+// make circular declaration possible.
+export interface List extends Array<Input> {}
+
+export interface Decoded {
+  data: Uint8Array | Uint8Array[]
+  remainder: Uint8Array
+}
 
 /**
  * RLP Encoding based on https://eth.wiki/en/fundamentals/rlp
- * This function takes in a data, convert it to Uint8Array if not, and a length for recursion
- * @param input - will be converted to Uint8Array
- * @returns returns Uint8Array of encoded data
+ * This function takes in data, converts it to Uint8Array if not,
+ * and adds a length for recursion.
+ * @param input Will be converted to Uint8Array
+ * @returns Uint8Array of encoded data
  **/
 export function encode(input: Input): Uint8Array {
   if (Array.isArray(input)) {
@@ -36,7 +34,7 @@ export function encode(input: Input): Uint8Array {
 
 /**
  * Slices a Uint8Array, throws if the slice goes out-of-bounds of the Uint8Array.
- * E.g. `safeSlice(RLP.utils.hexToBytes('aa'), 1, 2)` will throw.
+ * E.g. `safeSlice(hexToBytes('aa'), 1, 2)` will throw.
  * @param input
  * @param start
  * @param end
@@ -73,9 +71,9 @@ function encodeLength(len: number, offset: number): Uint8Array {
 
 /**
  * RLP Decoding based on https://eth.wiki/en/fundamentals/rlp
- * @param input - will be converted to Uint8Array
- * @param stream - Is the input a stream (false by default)
- * @returns - returns decode Array of Uint8Arrays containing the original message
+ * @param input Will be converted to Uint8Array
+ * @param stream Is the input a stream (false by default)
+ * @returns decoded Array of Uint8Arrays containing the original message
  **/
 export function decode(input: Uint8Array, stream?: boolean): Uint8Array
 export function decode(input: Uint8Array[], stream?: boolean): Uint8Array[]
@@ -93,7 +91,7 @@ export function decode(input: Input, stream = false): Uint8Array[] | Uint8Array 
     return decoded
   }
   if (decoded.remainder.length !== 0) {
-    throw new Error('invalid remainder')
+    throw new Error('invalid RLP: remainder must be zero')
   }
 
   return decoded.data
@@ -189,8 +187,112 @@ function _decode(input: Uint8Array): Decoded {
   }
 }
 
+const cachedHexes = Array.from({ length: 256 }, (_v, i) => i.toString(16).padStart(2, '0'))
+function bytesToHex(uint8a: Uint8Array): string {
+  // Pre-caching chars with `cachedHexes` speeds this up 6x
+  let hex = ''
+  for (let i = 0; i < uint8a.length; i++) {
+    hex += cachedHexes[uint8a[i]]
+  }
+  return hex
+}
+
+function parseHexByte(hexByte: string): number {
+  if (hexByte.length !== 2) throw new Error('Invalid byte sequence')
+  const byte = Number.parseInt(hexByte, 16)
+  if (Number.isNaN(byte)) throw new Error('Invalid byte sequence')
+  return byte
+}
+
+// Caching slows it down 2-3x
+function hexToBytes(hex: string): Uint8Array {
+  if (typeof hex !== 'string') {
+    throw new TypeError('hexToBytes: expected string, got ' + typeof hex)
+  }
+  if (hex.length % 2) throw new Error('hexToBytes: received invalid unpadded hex')
+  const array = new Uint8Array(hex.length / 2)
+  for (let i = 0; i < array.length; i++) {
+    const j = i * 2
+    array[i] = parseHexByte(hex.slice(j, j + 2))
+  }
+  return array
+}
+
+/** Concatenates two Uint8Arrays into one. */
+function concatBytes(...arrays: Uint8Array[]): Uint8Array {
+  if (arrays.length === 1) return arrays[0]
+  const length = arrays.reduce((a, arr) => a + arr.length, 0)
+  const result = new Uint8Array(length)
+  for (let i = 0, pad = 0; i < arrays.length; i++) {
+    const arr = arrays[i]
+    result.set(arr, pad)
+    pad += arr.length
+  }
+  return result
+}
+
+// Global symbols in both browsers and Node.js since v11
+// See https://github.com/microsoft/TypeScript/issues/31535
+declare const TextEncoder: any
+declare const TextDecoder: any
+
+function utf8ToBytes(utf: string): Uint8Array {
+  return new TextEncoder().encode(utf)
+}
+
+/** Transform an integer into its hexadecimal value */
+function numberToHex(integer: number | bigint): string {
+  if (integer < 0) {
+    throw new Error('Invalid integer as argument, must be unsigned!')
+  }
+  const hex = integer.toString(16)
+  return hex.length % 2 ? `0${hex}` : hex
+}
+
+/** Pad a string to be even */
+function padToEven(a: string): string {
+  return a.length % 2 ? `0${a}` : a
+}
+
+/** Check if a string is prefixed by 0x */
+function isHexPrefixed(str: string): boolean {
+  return str.length >= 2 && str[0] === '0' && str[1] === 'x'
+}
+
+/** Removes 0x from a given String */
+function stripHexPrefix(str: string): string {
+  if (typeof str !== 'string') {
+    return str
+  }
+  return isHexPrefixed(str) ? str.slice(2) : str
+}
+
+/** Transform anything into a Uint8Array */
+function toBytes(v: Input): Uint8Array {
+  if (v instanceof Uint8Array) {
+    return v
+  }
+  if (typeof v === 'string') {
+    if (isHexPrefixed(v)) {
+      return hexToBytes(padToEven(stripHexPrefix(v)))
+    }
+    return utf8ToBytes(v)
+  }
+  if (typeof v === 'number' || typeof v === 'bigint') {
+    if (!v) {
+      return Uint8Array.from([])
+    }
+    return hexToBytes(numberToHex(v))
+  }
+  if (v === null || v === undefined) {
+    return Uint8Array.from([])
+  }
+  throw new Error('toBytes: received unsupported type')
+}
+
 export const utils = {
   bytesToHex,
+  concatBytes,
   hexToBytes,
   utf8ToBytes,
 }
